+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/combat/fletching/entity/projectile/QuartzArrow.java
@@ -1,7 +1,9 @@
 package insane96mcp.iguanatweaksexpanded.module.combat.fletching.entity.projectile;
 
+import insane96mcp.iguanatweaksexpanded.module.combat.fletching.Fletching;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.projectile.Arrow;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
 public class QuartzArrow extends ITEArrow {
@@ -9,5 +11,10 @@ public class QuartzArrow extends ITEArrow {
         super(pEntityType, pLevel);
         this.friction = 0.995f;
         this.velocityMultiplier = 1.25f;
+    }
+
+    @Override
+    protected ItemStack getPickupItem() {
+        return new ItemStack(Fletching.QUARTZ_ARROW_ITEM.get());
     }
 }