+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * RIP fertility on server
+* pICKup arrows broken
 
 ## Upcoming
 * Added support for the Time Control mod
@@ -8,6 +9,7 @@
 * Expanded is now compatible with Pickaxes and shovels
 * Respawn obelisks now have a slightly lower chance to have iron blocks
 * You can now change your old respawn point when you have a respawn obelisk set
+* Fixed new arrows giving normal arrows when picked up
 
 ## 1.1.3
 * Reduced mobs attack reach by ~15%