+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Shields recipes are now consistent with other tools
+
 ## 1.1.4
 * Added support for the Time Control mod
   * If the mod is installed, the day and night durations change based off the current season