+++ changelog.md
@@ -5,6 +5,8 @@
 * Crate slowdown is now exponential (15% per crate above max -> 5%/20%/45%/80% for 1/2/3/4 crates above max)
 * Added Quaron stuff to tags
 * Corail woodcutter support has been replaced with Sawmill
+* Solarium and keego hoe cooldown reductions are now lower 
+* Slightly changed hammers cooldowns
 
 ## 1.1.4
 * Added support for the Time Control mod